add unit test for Akenie test handlers

+++ src/applications/handlers/comment/remove-comment.handler.ts
@@ -1,9 +1,10 @@
 import { RemoveCommentCommand } from '#applications/commands/comment/remove-comment.command';
 import { CommentRepository } from '#domain/repository/comment.repository';
 import { ICommandHandler } from '@nestjs/cqrs';
-import { Inject } from '@nestjs/common';
+import { Inject, NotFoundException } from '@nestjs/common';
 import { COMMENT_REPOSITORY } from '#shared/constantes/inject-token';
-
+import { CommandHandler } from '@nestjs/cqrs';
+@CommandHandler(RemoveCommentCommand)
 export class RemoveCommentHandler
     implements ICommandHandler<RemoveCommentCommand>
 {
@@ -12,6 +13,14 @@ export class RemoveCommentHandler
         private readonly commentRepository: CommentRepository
     ) {}
     async execute(command: RemoveCommentCommand) {
+        const comment = await this.commentRepository.getCommentsById(
+            command.commentId
+        );
+        if (!comment) {
+            throw new NotFoundException({
+                message: 'Comment non  Found',
+            });
+        }
         return this.commentRepository.removeComment(command.commentId);
     }
 }
